Update overview.rst
what is the goal...: ("") added to show quote.
upcoming is a better adj for update due to show that is a planned update.(for whole article)
Upcoming update lifecycle : "the" added several times for this reasons below:
-a word repeatedly used
-it was only word could be placed in context.
- a world specifically explained.
the verb "pass" refers to the noun " frame work" which is a singular one. so there is a "s" missing at the end.
there was a  extra "," after the world violated.
the conditional sentence can both be used in simple future and present.

number "2" is replaced.
"are right for" is replaced  with "match" to simplify the sentence and better content.